extract setting available keys to kit factories
