+++ src/kit.ts
@@ -25,5 +25,30 @@ const start = (audioBuffer: AudioBuffer) => {
   source.start();
 };
 
-const createKeyMap = (sampleBuffers: AudioBuffer[]) =>
+export const createKeyMap = (sampleBuffers: AudioBuffer[]) =>
   Object.fromEntries(sampleBuffers.map((sound, i) => [keys[i], sound]));
+
+/**
+ *
+ * @param slots Number of pads that will be visible to user
+ * @param octave The starting octave number on a MIDI device
+ */
+export const setAvailableKeys = (slots: number, octave: number): string[] => {
+  let renderedKeys = [];
+
+  const renderKeys = (slotsLeft: number) => {
+    if (slotsLeft > baseKeys.length) {
+      for (let i = 0; i < baseKeys.length; i++) {
+        renderedKeys.push(baseKeys[i] + String(octave));
+      }
+      octave++;
+      renderKeys(slotsLeft - baseKeys.length);
+    } else {
+      for (let i = 0; i < slotsLeft; i++) {
+        renderedKeys.push(baseKeys[i] + String(octave));
+      }
+    }
+  };
+  renderKeys(slots);
+  return renderedKeys;
+};
